+++ src/internal/modules/notifications-reports/lib/message-mapper.js
@@ -5,11 +5,11 @@ const titleCase = require('title-case');
 const isErrorStatus = status => status === 'error';
 
 /**
-   * Maps the displayStatus from the water service to a badge
-   * object which will render via Nunjucks
-   *
-   * @param {String} status
-   */
+ * Maps the displayStatus from the water service to a badge
+ * object which will render via Nunjucks
+ *
+ * @param {String} status
+ */
 const messageStatusBadgeMapper = status => {
   const text = titleCase(status);
   return isErrorStatus(status) ? { text, status } : { text };
@@ -20,49 +20,4 @@ const mapMessage = message => ({
   badge: messageStatusBadgeMapper(message.displayStatus)
 });
 
-const checkBoxItems = (filter, notificationCategories) => {
-  if (!filter) {
-    filter = [];
-  }
-  if (typeof filter === 'string') {
-    filter = [filter];
-  }
-  const returnsData = [];
-  if (notificationCategories) {
-    notificationCategories.forEach(category => {
-      returnsData.push({ value: category.value, text: category.label });
-    });
-  }
-
-  return returnsData.map(row => {
-    return {
-      ...row,
-      checked: filter.includes(row.value)
-    };
-  });
-};
-
-const mapResponseToView = (response, request, notificationCategories, sentBy = '') => {
-  let { filter } = request.query;
-  let filterString = '&';
-  if (filter) {
-    // Only one item selected
-    if (typeof filter === 'string') {
-      filter = [filter];
-    }
-    filter.forEach(param => {
-      filterString += `filter=${param}&`;
-    });
-  }
-  filterString += `sentBy=${sentBy}`;
-  filterString = filterString.replace('undefined', '');
-  return {
-    ...response,
-    filter,
-    sentBy,
-    checkBoxItems: checkBoxItems(filter, notificationCategories),
-    filterString
-  };
-};
-exports.mapResponseToView = mapResponseToView;
 exports.mapMessage = mapMessage;
